curl: make the URL indexes 64 bit
Otherwise we could misbehave already at 2 billion URLs and we can't have
that. A few of the counters are already correctly using the right type.

Closes #18096

+++ src/tool_operate.c
@@ -1250,7 +1250,7 @@ static CURLcode single_transfer(struct OperationConfig *config,
       }
       per->config = config;
       per->curl = curl;
-      per->urlnum = (unsigned int)u->num;
+      per->urlnum = u->num;
 
       /* default headers output stream is stdout */
       heads = &per->heads;
